feat: dashdash out; yargs in.
You now run it like `opsbot path/to/config.js`
